requirements.txt
pyyaml>=6.0.1
openai>=1.0.0
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
trl>=0.7.2
